Fix: Only one class per file
