+++ composer.json
@@ -15,7 +15,9 @@
         }
     ],
     "autoload": {
-        "classmap": ["lessc.inc.php"]
+        "classmap": [
+            "src/"
+        ]
     },
     "extra": {
         "branch-alias": {
