+++ lessify
@@ -1,6 +1,8 @@
 #!/usr/bin/php
 <?php
 
+require_once __DIR__ . '/vendor/autoload.php';
+
 if (php_sapi_name() != "cli") {
 	err($fa.$argv[0]." must be run in the command line.");
 	exit(1);
@@ -10,8 +12,6 @@ $exe = array_shift($argv); // remove filename
 if (!$fname = array_shift($argv)) {
 	exit("Usage: ".$exe." input-file\n");
 }
-
-require "lessify.inc.php";
 
 try  {
 	$parser = new lessify($fname);
