+++ src/lessc_formatter_lessjs.php
@@ -1,0 +1,8 @@
+<?php
+
+class lessc_formatter_lessjs extends lessc_formatter_classic {
+    public $disableSingle = true;
+    public $breakSelectors = true;
+    public $assignSeparator = ": ";
+    public $selectorSeparator = ",";
+}
