+++ tests/InputTest.php
@@ -1,7 +1,5 @@
 <?php
 
-require_once __DIR__ . "/../lessc.inc.php";
-
 // Runs all the tests in inputs/ and compares their output to ouputs/
 
 function _dump($value) {
