+++ tests/sort.php
@@ -1,8 +1,6 @@
 <?php
 error_reporting(E_ALL);
 
-require realpath(dirname(__FILE__)).'/../lessc.inc.php';
-
 // sorts the selectors in stylesheet in order to normalize it for comparison
 
 $exe = array_shift($argv); // remove filename
